requirements.txt
gradio
groq
python-dotenv
PyPDF2
reportlab
